backend/requirements.txt
Django==4.2
djangorestframework==3.14.0
pandas==1.5.3
reportlab==4.0.5
django-cors-headers==4.3.1
